fixed multiple warnings and wrong usage of signed integer

+++ torlib/Cell.cpp
@@ -66,7 +66,7 @@ void Cell::Append(string& val) {
 	index += val.length();
 }
 
-void Cell::Append(unc* segment, int length) {
+void Cell::Append(unc* segment, size_t length) {
 	memcpy(buffer + index, segment, length);
 	index += length;
 }
@@ -92,7 +92,7 @@ unsigned char* Cell::GetPayload() {
 	return buffer + POSITION_PAYLOAD_SIZE;
 }
 
-int Cell::GetPayloadSize() {
+size_t Cell::GetPayloadSize() {
 	return sizeof(buffer) - POSITION_PAYLOAD_SIZE;
 }
 
@@ -100,12 +100,12 @@ unsigned char* Cell::GetBuffer() {
 	return buffer;
 }
 
-int Cell::GetBufferSize() {
+size_t Cell::GetBufferSize() {
 	if (IsVariableLen())
 		return sizeof(buffer);
 	else return CELL_SIZE;
 }
-int Cell::GetCurentIndex() {
+size_t Cell::GetCurentIndex() {
 	return index;
 }
